# book-network/src/main/java/com/devskills/book/book/BookResponse.java
package com.devskills.book.book;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class BookResponse {
	
	private Integer id;
	private String title;
	private String authorName;
	private String isbn;
	private String synopsis;
	private String owner;
	private byte[] cover;
	private double rate;
	private boolean archived;
	private boolean shareable;

}